+++ test/testImage.php
@@ -6,14 +6,15 @@ require 'autoload.php';
 
 $config = require '../src/config.php';
 
+$type = $_GET['type'] ?? 'background'; // 获取 type 参数，默认为 background
 
-function showBlock()
+function showBlock($type)
 {
     global $config;
     $factory = new Factory($config);
     $blockImage = $factory->makeBlockImage();
     $blockImage->run();
-    $blockImage->echo();
+    $blockImage->echo($type);
 }
 
 function showWord()
@@ -25,4 +26,9 @@ function showWord()
     $blockImage->echo();
 }
 
-showBlock();
+// 简单的路由判断
+if (isset($_GET['mode']) && $_GET['mode'] === 'word') {
+    showWord();
+} else {
+    showBlock($type);
+}